requirements.txt
Django==4.2.21
django-bootstrap5==25.1
django-simple-history==3.8.0
django-leaflet==0.32.0
django-tables2==2.7.5
django-debug-toolbar==5.2.0
django-filter==25.1
django-crispy-forms==2.4
crispy-bootstrap5==2025.4
requests==2.32.3
openpyxl==3.1.5
mappyfile==1.0.2
mapscript==7.6.0
slugify==0.0.1
python-magic==0.4.27